+++ src/data_structures/map.py
@@ -3,7 +3,7 @@
 from enum import Enum
 from PIL.Image import Image
 from src.data_structures.field import Field, FieldType, FieldLike, MapField, FieldNullError
-from src.utilities.utilities import colour_wheel
+from src.utilities.utils import colour_wheel
 
 
 class MapType(Enum):
