+++ src/scripts/materials_add.py
@@ -45,7 +45,7 @@ def decode_lattice_type(value):
 
 def add(database):
 	print()
-	IDs = utilities.utilities.parse_ids(input('Enter material IDs to add separated by commas/hyphens: '))
+	IDs = utilities.utils.parse_ids(input('Enter material IDs to add separated by commas/hyphens: '))
 	materials = Phase.load_all(Config(CONFIG_PATH).project.phase_dir)
 
 	for ID in IDs:
